putting the sort logic in AllQuotes; fixing url issue when the drag/drop order stays the same
